ml/requirements.txt
# ── 深度学习框架 ───────────────────────────────────────────────────────────
# CUDA 12.x 用户请先安装对应版本的 PyTorch，例如：
# pip install torch --index-url https://download.pytorch.org/whl/cu121
torch>=2.2.0

# ── Hugging Face 生态 ──────────────────────────────────────────────────────
transformers>=4.45.0
datasets>=2.20.0
peft>=0.12.0
trl>=0.11.0
accelerate>=0.33.0
tokenizers>=0.20.0
safetensors>=0.4.0

# ── 4-bit / 8-bit 量化（可选，仅 CUDA 支持）──────────────────────────────
bitsandbytes>=0.43.0; sys_platform != "win32"
# Windows 用户安装预编译包：
# pip install bitsandbytes --index-url https://github.com/jllllll/bitsandbytes-windows-webui/releases/download/wheels

# ── 推理服务器 ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0

# ── 配置文件 ───────────────────────────────────────────────────────────────
pyyaml>=6.0

# ── 其他工具 ───────────────────────────────────────────────────────────────
tqdm>=4.66.0
numpy>=1.26.0
